requirements.txt
pypdf
reportlab
python-docx